Port to .NET Standard

+++ TmpEmulation/TmpIO/TmpAtlasPixel.cs
@@ -5,7 +5,8 @@ namespace TmpIO;
 [StructLayout(LayoutKind.Sequential, Pack = 1)]
 public struct TmpAtlasPixel
 {
-    public Half R;
-    public Half G;
-    public Half B;
+    // Half-precision floating-point format (16 bits per channel)
+    public ushort R;
+    public ushort G;
+    public ushort B;
 }
